Polish selection cleanup follow-ups
Agent-Logs-Url: https://github.com/starseeker/brlcad_quickiterate/sessions/ca34be51-cad2-4c8a-b977-dc292bc7da9c

+++ brlcad/src/libged/view/view.c
@@ -411,7 +411,8 @@ _view_cmd_selections(void *bs, int argc, const char **argv)
 	    sel_cnt = bsg_selection_count(ss);
     }
 
-    bu_vls_printf(gd->gedp->ged_result_str, "%zu", sel_cnt);
+    bu_vls_printf(gd->gedp->ged_result_str, "%llu",
+	    (unsigned long long)sel_cnt);
 
     return BRLCAD_OK;
 }
